refactor: auth 인증 로직을 service/store/support 패키지로 분리
