+++ Pokade/src/main/java/com/pokade/domain/auth/store/PasswordResetCodeStore.java
@@ -1,4 +1,4 @@
-package com.pokade.domain.auth.service;
+package com.pokade.domain.auth.store;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.StringRedisTemplate;
