+++ Pokade/src/test/java/com/pokade/domain/auth/service/VerificationCodeGeneratorTest.java
@@ -1,5 +1,6 @@
 package com.pokade.domain.auth.service;
 
+import com.pokade.domain.auth.support.VerificationCodeGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
